prima parte/weather_notifier/requirements.txt
requests
python-telegram-bot==20.7
psycopg2